ref: Common code base for release files endpoint
