+++ src/sentry/api/endpoints/project_release_files.py
@@ -6,43 +6,39 @@ from django.db.models import Q
 from rest_framework.response import Response
 
 from sentry.api.bases.project import ProjectEndpoint, ProjectReleasePermission
-from sentry.api.endpoints.organization_release_files import load_dist
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.paginator import OffsetPaginator
 from sentry.api.serializers import serialize
 from sentry.constants import MAX_RELEASE_FILES_OFFSET
-from sentry.models import File, Release, ReleaseFile
+from sentry.models import Distribution, File, Release, ReleaseFile
 
 ERR_FILE_EXISTS = "A file matching this name already exists for the given release"
 _filename_re = re.compile(r"[\n\t\r\f\v\\]")
 
 
-class ProjectReleaseFilesEndpoint(ProjectEndpoint):
-    permission_classes = (ProjectReleasePermission,)
+def load_dist(results):
+    # Dists are pretty uncommon.  In case they do appear load them now
+    # as trying to join this on the DB does terrible things with large
+    # offsets (it would otherwise generate a left outer join).
+    dists = dict.fromkeys(x.dist_id for x in results)
+    if not dists:
+        return results
 
-    def get(self, request, project, version):
-        """
-        List a Project Release's Files
-        ``````````````````````````````
+    for dist in Distribution.objects.filter(pk__in=dists.keys()):
+        dists[dist.id] = dist
 
-        Retrieve a list of files for a given release.
+    for result in results:
+        if result.dist_id is not None:
+            dist = dists.get(result.dist_id)
+            if dist is not None:
+                result.dist = dist
 
-        :pparam string organization_slug: the slug of the organization the
-                                          release belongs to.
-        :pparam string project_slug: the slug of the project to list the
-                                     release files of.
-        :pparam string version: the version identifier of the release.
-        :qparam string query: If set, this parameter is used to search files.
-        :auth: required
-        """
+    return results
+
+
+class ReleaseFilesMixin:
+    def get_releasefiles(self, request, release):
         query = request.GET.getlist("query")
-
-        try:
-            release = Release.objects.get(
-                organization_id=project.organization_id, projects=project, version=version
-            )
-        except Release.DoesNotExist:
-            raise ResourceDoesNotExist
 
         file_list = (
             ReleaseFile.public_objects.filter(release=release)
@@ -68,45 +64,8 @@ class ProjectReleaseFilesEndpoint(ProjectEndpoint):
             on_results=lambda r: serialize(load_dist(r), request.user),
         )
 
-    def post(self, request, project, version):
-        """
-        Upload a New Project Release File
-        `````````````````````````````````
-
-        Upload a new file for the given release.
-
-        Unlike other API requests, files must be uploaded using the
-        traditional multipart/form-data content-type.
-
-        The optional 'name' attribute should reflect the absolute path
-        that this file will be referenced as. For example, in the case of
-        JavaScript you might specify the full web URI.
-
-        :pparam string organization_slug: the slug of the organization the
-                                          release belongs to.
-        :pparam string project_slug: the slug of the project to change the
-                                     release of.
-        :pparam string version: the version identifier of the release.
-        :param string name: the name (full path) of the file.
-        :param string dist: the name of the dist.
-        :param file file: the multipart encoded file.
-        :param string header: this parameter can be supplied multiple times
-                              to attach headers to the file.  Each header
-                              is a string in the format ``key:value``.  For
-                              instance it can be used to define a content
-                              type.
-        :auth: required
-        """
-        try:
-            release = Release.objects.get(
-                organization_id=project.organization_id, projects=project, version=version
-            )
-        except Release.DoesNotExist:
-            raise ResourceDoesNotExist
-
-        logger = logging.getLogger("sentry.files")
-        logger.info("projectreleasefile.start")
-
+    @staticmethod
+    def post_releasefile(request, release, logger):
         if "file" not in request.data:
             return Response({"detail": "Missing uploaded file"}, status=400)
 
@@ -169,3 +128,72 @@ class ProjectReleaseFilesEndpoint(ProjectEndpoint):
             return Response({"detail": ERR_FILE_EXISTS}, status=409)
 
         return Response(serialize(releasefile, request.user), status=201)
+
+
+class ProjectReleaseFilesEndpoint(ProjectEndpoint, ReleaseFilesMixin):
+    permission_classes = (ProjectReleasePermission,)
+
+    def get(self, request, project, version):
+        """
+        List a Project Release's Files
+        ``````````````````````````````
+
+        Retrieve a list of files for a given release.
+
+        :pparam string organization_slug: the slug of the organization the
+                                          release belongs to.
+        :pparam string project_slug: the slug of the project to list the
+                                     release files of.
+        :pparam string version: the version identifier of the release.
+        :qparam string query: If set, this parameter is used to search files.
+        :auth: required
+        """
+        try:
+            release = Release.objects.get(
+                organization_id=project.organization_id, projects=project, version=version
+            )
+        except Release.DoesNotExist:
+            raise ResourceDoesNotExist
+
+        return self.get_releasefiles(request, release)
+
+    def post(self, request, project, version):
+        """
+        Upload a New Project Release File
+        `````````````````````````````````
+
+        Upload a new file for the given release.
+
+        Unlike other API requests, files must be uploaded using the
+        traditional multipart/form-data content-type.
+
+        The optional 'name' attribute should reflect the absolute path
+        that this file will be referenced as. For example, in the case of
+        JavaScript you might specify the full web URI.
+
+        :pparam string organization_slug: the slug of the organization the
+                                          release belongs to.
+        :pparam string project_slug: the slug of the project to change the
+                                     release of.
+        :pparam string version: the version identifier of the release.
+        :param string name: the name (full path) of the file.
+        :param string dist: the name of the dist.
+        :param file file: the multipart encoded file.
+        :param string header: this parameter can be supplied multiple times
+                              to attach headers to the file.  Each header
+                              is a string in the format ``key:value``.  For
+                              instance it can be used to define a content
+                              type.
+        :auth: required
+        """
+        try:
+            release = Release.objects.get(
+                organization_id=project.organization_id, projects=project, version=version
+            )
+        except Release.DoesNotExist:
+            raise ResourceDoesNotExist
+
+        logger = logging.getLogger("sentry.files")
+        logger.info("projectreleasefile.start")
+
+        return self.post_releasefile(request, release, logger)
